Reverted color arithmatic to float from double

+++ heavenLi_pyopengl/drawUtils/drawEllipse.cpp
@@ -9,11 +9,11 @@ int drawEllipse(
       float bsx,                 /* x-Scale 2.0=spans display before GL scaling */
       float bsy,                 /* y-Scale 2.0=spans display before GL scaling */
       char circleSegments,       /* Number of sides */
-      double *color,             /* Polygon Color */
+      float *color,             /* Polygon Color */
       std::vector<float> &verts, /* Input Vector of x,y coordinates */
-      std::vector<double> &colrs  /* Input Vector of r,g,b values */
+      std::vector<float> &colrs  /* Input Vector of r,g,b values */
       ){
-   double R, G, B;
+   float R, G, B;
    char degSegment = 360 / circleSegments;
    R = color[0];
    G = color[1];
@@ -43,9 +43,9 @@ int drawEllipse(
       float by,                  /* Y-Coordinate */
       float bs,                  /* Scale~ 2.0=spans display before GL Transformations */
       char circleSegments,       /* Number of sides */
-      double *color,             /* Polygon Color */
+      float *color,             /* Polygon Color */
       std::vector<float> &verts, /* Input Vector of x,y coordinates */
-      std::vector<double> &colrs  /* Input Vector of r,g,b values */
+      std::vector<float> &colrs  /* Input Vector of r,g,b values */
       ){
    return drawEllipse(bx, by, bs, bs, circleSegments, color, verts, colrs);
 }
@@ -56,9 +56,9 @@ int drawCircle(
       float by,                  /* Y-Coordinate */
       float bs,                  /* Scale~ 2.0=spans display before GL Transformations */
       char circleSegments,       /* Number of sides */
-      double *color,             /* Polygon Color */
+      float *color,             /* Polygon Color */
       std::vector<float> &verts, /* Input Vector of x,y coordinates */
-      std::vector<double> &colrs  /* Input Vector of r,g,b values */
+      std::vector<float> &colrs  /* Input Vector of r,g,b values */
       ){
    return drawEllipse(bx, by, bs, bs, circleSegments, color, verts, colrs);
 }
