+++ heavenLi_pyopengl/drawUtils/rgb2hsv.cpp
@@ -2,8 +2,8 @@
 
 using namespace std;
 
-void rgb2hsv(double r, double g, double b, double hsv[3]) {
-   double min, max, delta;
+void rgb2hsv(float r, float g, float b, float hsv[3]) {
+   float min, max, delta;
 
    min = r   < g ? r   : g;
    min = min < b ? min : b;
@@ -29,16 +29,16 @@ void rgb2hsv(double r, double g, double b, double hsv[3]) {
    }
 
    if ( r >= max ) {
-      hsv[0] = 0.0 + ( g - b ) / delta;
+      hsv[0] = 0.0f + ( g - b ) / delta;
    } else if ( g >= max ) {
-      hsv[0] = 2.0 + ( b - r ) / delta;
+      hsv[0] = 2.0f + ( b - r ) / delta;
    } else {
-      hsv[0] = 4.0 + ( r - g ) / delta;
+      hsv[0] = 4.0f + ( r - g ) / delta;
    }
 
-   hsv[0] *= 60.0 / 360.0;
-   if ( hsv[0] < 0.0 )
-      hsv[0] += 1.0;
+   hsv[0] *= 60.0f / 360.0f;
+   if ( hsv[0] < 0.0f )
+      hsv[0] += 1.0f;
 
    return;
 }
